+++ examples/todos/src/ui.rs
@@ -14,17 +14,18 @@ impl Compose for Todos {
           $this.write().new_task(text.to_string());
         })
         @Tabs {
+          h_align: HAlign::Stretch,
           @Tab {
-            @TabItem { @Label::new("ALL") }
-            @TabPane(task_lists(this.clone_writer(), |_| true))
+            label: "All",
+            @ task_lists(this.clone_writer(), |_| true)
           }
           @Tab {
-            @TabItem { @{ Label::new("ACTIVE") } }
-            @TabPane(task_lists(this.clone_writer(), |t| !t.complete ))
+            label: "ACTIVE",
+            @ task_lists(this.clone_writer(), |t| !t.complete )
           }
           @Tab {
-            @TabItem { @{ Label::new("DONE") } }
-            @TabPane(task_lists(this, |t| t.complete ))
+            label: "DONE",
+            @task_lists(this, |t| t.complete )
           }
         }
       }