+++ CHANGELOG.md
@@ -44,6 +44,7 @@ Please only add new entries below the [Unreleased](#unreleased---releasedate) he
 ### Changed
 
 - **widgets**: Refactor `Divider` Widget. (#702 @wjian23)
+- **widgets**: Refactor `Tabs` Widget. (#pr @M-Adoo)
 - **ci**: update rust version of ci to 2025-03-06 (#702 @wjian23)
 
 ### Breaking